Move target resolution to Target
